+++ engine/apps/slack/tests/test_scenario_steps/test_resolution_note.py
@@ -201,10 +201,7 @@ def test_resolution_notes_modal_closed_before_update(
 
     alert_receive_channel = make_alert_receive_channel(organization)
     alert_group = make_alert_group(alert_receive_channel)
-    slack_message = make_slack_message(
-        alert_group=alert_group, channel_id="RANDOM_CHANNEL_ID", slack_id="RANDOM_MESSAGE_ID"
-    )
-    slack_message.get_alert_group()  # fix FKs
+    make_slack_message(alert_group=alert_group, channel_id="RANDOM_CHANNEL_ID", slack_id="RANDOM_MESSAGE_ID")
 
     payload = {
         "trigger_id": "TEST",