+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Make Slack integration not post an alert group message if it's already deleted + refactor AlertGroup and
+  SlackMessage foreign key relationship by @vadimkerr ([#2957](https://github.com/grafana/oncall/pull/2957))
+
 ## v1.3.31 (2023-09-04)
 
 ### Fixed